make additional instrumentation part of base instruments instead of just trucustomapp.
