+++ trulens_eval/trulens_eval/instruments.py
@@ -1020,3 +1020,43 @@ class Instrument(object):
             logger.debug(
                 f"{query}: Do not know how to instrument object of type {cls}."
             )
+
+
+class instrument:
+    """
+    Decorator for marking methods to be instrumented in custom classes that are
+    wrapped by App.
+    """
+
+    # https://stackoverflow.com/questions/2366713/can-a-decorator-of-an-instance-method-access-the-class
+
+    def __init__(self, func: Callable):
+        self.func = func
+
+    def __set_name__(self, cls: type, name: str):
+        """
+        For use as method decorator.
+        """
+
+        # Important: do this first:
+        setattr(cls, name, self.func)
+
+        # Note that this does not actually change the method, just adds it to
+        # list of filters.
+        self.method(cls, name)
+
+    @classmethod
+    def method(self_class, cls: type, name: str) -> None:
+        # Add the class with a method named `name`, its module, and the method
+        # `name` to the Default instrumentation walk filters.
+        Instrument.Default.MODULES.add(cls.__module__)
+        Instrument.Default.CLASSES.add(cls)
+
+        check_o = Instrument.Default.METHODS.get(name, lambda o: False)
+        Instrument.Default.METHODS[
+            name] = lambda o: check_o(o) or isinstance(o, cls)
+
+    @classmethod
+    def methods(self_class, cls: type, names: Iterable[str]) -> None:
+        for name in names:
+            self_class.method(cls, name)
